Changed management of returns in WBlockExpressions.
Now, a WBlockExpression with returns is typed accordingly to the return
expressions. Only in absence of a return it is typed to the last
expression.

Closures inherit the type of its body (a WBlockExpression).
Methods only inherit the type when (a) there are returns inside the
method or (b) it is a compact method (using = syntax, no return
required). Otherwise, the method is void.